refactor: 移除不必要的类型转换，简化代码
- 移除 IdentityNumber.format() 方法调用中的显式类型转换
- 简化 SimpleIdentityCodec 中的类型转换
- 更新 README.md 中的示例代码，保持一致性
- 所有测试用例同步更新以匹配新的方法签名
- 代码更加简洁易读，减少冗余的类型声明

+++ src/main/java/io/github/nextentity/codec/identity/SimpleIdentityCodec.java
@@ -116,7 +116,7 @@ public class SimpleIdentityCodec implements IdentityCodec {
 
         // 6. 校验年份是否为四位数（年份应在 0000-9999 范围内）
         int year = birth.getYear();
-        return IdentityNumber.format(address, (short) year, (byte) birth.getMonthValue(), (byte) birth.getDayOfMonth(), (short) sequence);
+        return IdentityNumber.format(address, year, birth.getMonthValue(), birth.getDayOfMonth(), sequence);
     }
 
 }
